[NativeAOT] Inline array map anchors
Keep the shared array typemap anchor types in an existing compiled Mono.Android runtime source file so the project does not need a separate Compile item for ArrayMapAnchors.cs.
