+++ src/Mono.Android/Mono.Android.csproj
@@ -361,7 +361,6 @@
     <Compile Include="Java.Util.Concurrent.Atomic\AtomicLong.cs" />
     <Compile Include="Javax.Microedition.Khronos.Egl\EGLContext.cs" />
     <Compile Include="Microsoft.Android.Runtime\AggregateTypeMap.cs" />
-    <Compile Include="Microsoft.Android.Runtime\ArrayMapAnchors.cs" />
     <Compile Include="Microsoft.Android.Runtime\ITypeMapWithAliasing.cs" />
     <Compile Include="Microsoft.Android.Runtime\ManagedTypeManager.cs" />
     <Compile Include="Microsoft.Android.Runtime\ManagedTypeMapping.cs" />
